backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

# Core Framework
flask
Flask==3.0.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.1
Flask-Security-Too==5.4.3
python-dotenv==1.0.1
werkzeug==3.0.3
Flask-Limiter==3.8.0

# Database & Vector Support
psycopg2-binary==2.9.9
pgvector==0.3.2
SQLAlchemy==2.0.30
numpy==1.26.4

requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2

# AI Interview (Groq)
groq

# AWS S3 Storage
boto3>=1.34.0

# Production
gunicorn==22.0.0
feedparser==6.0.11
